requirements.txt
annotated-types==0.7.0
anyio==4.12.1
bcrypt==4.0.1
certifi==2026.1.4
click==8.3.1
colorama==0.4.6
distro==1.9.0
ecdsa==0.19.1
fastapi==0.115.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
jiter==0.13.0
openai==2.23.0
passlib==1.7.4
pyasn1==0.6.2
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.0.1
python-jose==3.5.0
python-multipart==0.0.12
PyYAML==6.0.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.38.6
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.32.0
watchfiles==1.1.1
websockets==16.0
pg8000==1.31.2
aiofiles==23.2.1
